tmp/spaghetti/requirements_docs.txt
sphinx>=1.4.3
sphinxcontrib-bibtex
sphinx_bootstrap_theme
numpydoc
esda